+++ forms/jqm/elements/choiceexpanded.js
@@ -43,7 +43,6 @@ define(function (require) {
       this._renderOptions();
 
       if (type === 'select') {
-        this.bindRivets();
         this.model.on('change:value', this.onSelectValueChange, this);
         this.onSelectValueChange();
       } else { // type === 'multi'
@@ -150,6 +149,8 @@ define(function (require) {
       if (_.difference(value, _.keys(model.attributes.options)).length > 0) {
         view.$el.find('input[type = text]').val(_.difference(value, _.keys(model.attributes.options)));
       }
+      this.model.isValid();
+      this.renderErrors(this.model);
     },
 
     onSelectValueChange: function () {
@@ -179,6 +180,8 @@ define(function (require) {
         // Also need to fill the text box back in, in addition to selecting radio
         this.$el.find('input[type = text]').val(this.model.get('value'));
       }
+      this.model.isValid();
+      this.renderErrors(this.model);
     },
 
     fetchValue: function () {