+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # Changelog
 
 
+## v3.4.1 - 2015-07-28
+
+
+### Fixed
+
+- FORMS-203: improve Forms input latency with `_.throttle()`
+
+    - also drop Rivets dependency and explicitly bind event
+
+- FORMS-206: required choice fields that and not empty no longer cause
+  validation errors
+
+
 ## v3.4.0 - 2015-07-16
 
 
@@ -28,7 +41,7 @@
 
 ### Fixed
 
-- FORMS-201: native date picker displays "now" and "now +" values
+- FORMS-189: native date picker displays "now" and "now +" values
 
 
 ## v3.3.5 - 2015-06-26